requirements.txt
flask>=3.0
pyyaml>=6.0
requests>=2.31
beautifulsoup4>=4.12
openai>=1.0
anthropic>=0.20
reportlab>=4.0
pillow>=10.0
playwright>=1.40
